多个经典CNN网络实现FashionMNIST分类/requirements.txt
python==3.8
 Pytorch==1.10.1
 Cudatookit==11.3
 Cudnn==8.2
 torchsummary==1.5.1
 numpy==1.23.2
 pandas==1.3.4
 matplotlib==3.5.0
 sklearn==0.0